patch to make Bpod save the photometry plot figures
eventually InitializeNidaq change needs to be rewritten so that instead of outputting outputs, it just assigns them to BpodSystem instead.  for now, I put the BpodSystem writing at the end so that the other changes that I made for figure saving will work.
added a Photometry-specific saving code based on the savesessionfig code and added a call for the photometrysave to RunProtocol